+++ lib/determinator/control.rb
@@ -27,6 +27,7 @@ module Determinator
     # @param :id [#to_s] The id of the actor being determinated for
     # @param :guid [#to_s] The Anonymous id of the actor being determinated for
     # @param :properties [Hash<Symbol,String>] The properties of this actor which will be used for including this actor or not
+    # @raise [ArgumentError] When the arguments given to this method aren't ever going to produce a useful response
     # @return [true,false] Whether the feature is on (true) or off (false) for this actor
     def feature_flag_on?(name, id: nil, guid: nil, properties: {})
       determinate(name, id: id, guid: guid, properties: properties) do |feature|
@@ -40,6 +41,7 @@ module Determinator
     # @param :id [#to_s] The id of the actor being determinated for
     # @param :guid [#to_s] The Anonymous id of the actor being determinated for
     # @param :properties [Hash<Symbol,String>] The properties of this actor which will be used for including this actor or not
+    # @raise [ArgumentError] When the arguments given to this method aren't ever going to produce a useful response
     # @return [false,String] Returns false, if the actor is not in this experiment, or otherwise the variant name.
     def which_variant(name, id: nil, guid: nil, properties: {})
       determinate(name, id: id, guid: guid, properties: properties) do |feature|
@@ -58,23 +60,23 @@ module Determinator
     def determinate(name, id:, guid:, properties:)
       feature = retrieval.retrieve(name)
       if feature.nil?
-        Determinator.missing_feature(name)
+        Determinator.notice_missing_feature(name)
         return false
       end
 
       # Calling method can place constraints on the feature, eg. experiment only
       return false if block_given? && !yield(feature)
 
-      # Overrides take precedence
-      return feature.override_value_for(id) if feature.overridden_for?(id)
-
+      # Inactive features are always, always off
       return false unless feature.active?
+
+      return feature.override_value_for(id) if feature.overridden_for?(id)
 
       target_group = choose_target_group(feature, properties)
       # Given constraints have excluded this actor from this experiment
       return false unless target_group
 
-      indicators = indicators_for(feature, id, guid)
+      indicators = indicators_for(feature, actor_identifier(feature, id, guid))
       # This actor isn't described in enough detail to form indicators
       return false unless indicators
 
@@ -86,30 +88,53 @@ module Determinator
       return true unless feature.experiment?
 
       variant_for(feature, indicators.variant)
+
+    rescue ArgumentError
+      raise
+
+    rescue => e
+      Determinator.notice_error(e)
+      false
     end
 
     def choose_target_group(feature, properties)
-      # Keys must be strings
-      normalised_properties = properties.each_with_object({}) do |(k, v), h|
-        h[k.to_s] = v
+      # Keys and values must be strings
+      normalised_properties = properties.each_with_object({}) do |(name, values), hash|
+        hash[name.to_s] = [*values].map(&:to_s)
       end
 
       feature.target_groups.select { |tg|
+        next false unless tg.rollout.between?(1, 65_536)
+
         tg.constraints.reduce(true) do |fit, (scope, *required)|
-          present = [*normalised_properties[scope.to_s]]
+          present = [*normalised_properties[scope]]
           fit && (required.flatten & present.flatten).any?
         end
       # Must choose target group deterministically, if more than one match
       }.sort_by { |tg| tg.rollout }.last
     end
 
-    def indicators_for(feature, id, guid)
-      # If we're slicing by guid then we never pay attention to id
-      actor_identifier = case feature.bucket_type
-                         when :id       then id
-                         when :guid     then guid
-                         when :fallback then id || guid
-                         end
+    def actor_identifier(feature, id, guid)
+      case feature.bucket_type
+      when :id
+        id
+      when :guid
+        return guid if guid.to_s != ''
+
+        raise ArgumentError, 'A GUID must always be given for GUID bucketed features'
+      when :fallback
+        identifier = (id || guid).to_s
+        return identifier if identifier != ''
+
+        raise ArgumentError, 'An ID or GUID must always be given for Fallback bucketed features'
+      when :single
+        'all'
+      else
+        Determinator.notice_error "Cannot process the '#{feature.bucket_type}' bucket type found in #{feature.name}"
+      end
+    end
+
+    def indicators_for(feature, actor_identifier)
       # No identified means not enough info was given by the caller
       # to determine an outcome for this feature
       return unless actor_identifier
@@ -132,12 +157,12 @@ module Determinator
       variant_weight_total = feature.variants.values.reduce(:+)
       scale_factor = 65_535 / variant_weight_total.to_f
 
+      sorted_variants = feature.variants.keys.sort
       # Find the variant the indicator sits within
-      previous_upper_bound = 0
-      feature.variants.each do |name, weight|
-        new_upper_bound = previous_upper_bound + scale_factor * weight
-        return name if indicator <= new_upper_bound
-        previous_upper_bound = new_upper_bound
+      upper_bound = 0
+      sorted_variants.each do |variant_name|
+        upper_bound = upper_bound + scale_factor * feature.variants[variant_name]
+        return variant_name if indicator <= upper_bound
       end
 
       raise ArgumentError, "A variant should have been found by this point, there is a bug in the code."